# perf(turbo-tasks): Call `.shrink_to_fit()` on common collection types when constructing a cell