--- a/turbopack/crates/turbo-tasks-backend/tests/shrink_to_fit.rs
+++ b/turbopack/crates/turbo-tasks-backend/tests/shrink_to_fit.rs
@@ -1,0 +1,1 @@
+../../turbo-tasks-testing/tests/shrink_to_fit.rs