--- a/turbopack/crates/turbo-tasks-memory/tests/shrink_to_fit.rs
+++ b/turbopack/crates/turbo-tasks-memory/tests/shrink_to_fit.rs
@@ -1,0 +1,1 @@
+../../turbo-tasks-testing/tests/shrink_to_fit.rs