--- a/turbopack/crates/turbo-tasks-testing/tests/shrink_to_fit.rs
+++ b/turbopack/crates/turbo-tasks-testing/tests/shrink_to_fit.rs
@@ -1,5 +1,6 @@
 #![feature(arbitrary_self_types)]
 #![feature(arbitrary_self_types_pointers)]
+#![allow(clippy::needless_return)] // tokio macro-generated code doesn't respect this
 
 use anyhow::Result;
 use turbo_tasks::Vc;