--- a/turbopack/crates/turbo-tasks/src/macro_helpers.rs
+++ b/turbopack/crates/turbo-tasks/src/macro_helpers.rs
@@ -1,4 +1,6 @@
 //! Runtime helpers for [turbo-tasks-macro].
+use std::ops::{Deref, DerefMut};
+
 pub use async_trait::async_trait;
 pub use once_cell::sync::{Lazy, OnceCell};
 pub use serde;
@@ -9,8 +11,8 @@ pub use super::{
     manager::{find_cell_by_type, notify_scheduled_tasks, spawn_detached_for_testing},
 };
 use crate::{
-    debug::ValueDebugFormatString, task::TaskOutput, RawVc, ResolvedValue, TaskInput,
-    TaskPersistence, Vc,
+    debug::ValueDebugFormatString, shrink_to_fit::ShrinkToFit, task::TaskOutput, RawVc,
+    ResolvedValue, TaskInput, TaskPersistence, Vc,
 };
 
 #[inline(never)]
@@ -55,4 +57,61 @@ macro_rules! stringify_path {
     ($path:path) => {
         stringify!($path)
     };
+}
+
+/// A wrapper type that uses the [autoderef specialization hack][autoderef] to call
+/// [`ShrinkToFit::shrink_to_fit`] on types that implement [`ShrinkToFit`].
+///
+/// This uses a a no-op method [`ShrinkToFitFallbackNoop::shrink_to_fit`] on types that do not
+/// implement [`ShrinkToFit`].
+///
+/// This is used by the derive macro for [`ShrinkToFit`], which is called by the
+/// [turbo_tasks::value][crate::value] macro.
+///
+/// [autoderef]: http://lukaskalbertodt.github.io/2019/12/05/generalized-autoref-based-specialization.html
+pub struct ShrinkToFitDerefSpecialization<'a, T> {
+    inner: ShrinkToFitFallbackNoop<'a, T>,
+}
+
+impl<'a, T> ShrinkToFitDerefSpecialization<'a, T> {
+    pub fn new(real: &'a mut T) -> Self {
+        Self {
+            inner: ShrinkToFitFallbackNoop { real },
+        }
+    }
+}
+
+impl<T> ShrinkToFitDerefSpecialization<'_, T>
+where
+    T: ShrinkToFit,
+{
+    pub fn shrink_to_fit(&mut self) {
+        // call the real `ShrinkToFit::shrink_to_fit` method
+        self.inner.real.shrink_to_fit()
+    }
+}
+
+impl<'a, T> Deref for ShrinkToFitDerefSpecialization<'a, T> {
+    type Target = ShrinkToFitFallbackNoop<'a, T>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.inner
+    }
+}
+
+impl<T> DerefMut for ShrinkToFitDerefSpecialization<'_, T> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.inner
+    }
+}
+
+// Implements `ShrinkToFit` using a no-op `ShrinkToFit::shrink_to_fit` method.
+pub struct ShrinkToFitFallbackNoop<'a, T> {
+    real: &'a mut T,
+}
+
+impl<T> ShrinkToFitFallbackNoop<'_, T> {
+    /// A no-op function called as part of [`ShrinkToFitDerefSpecialization`] when `T` does not
+    /// implement [`ShrinkToFit`].
+    pub fn shrink_to_fit(&mut self) {}
 }