--- a/turbopack/crates/turbo-tasks/src/primitives.rs
+++ b/turbopack/crates/turbo-tasks/src/primitives.rs
@@ -8,11 +8,11 @@ use crate::{
 };
 
 __turbo_tasks_internal_primitive!(());
-__turbo_tasks_internal_primitive!(String);
+__turbo_tasks_internal_primitive!(String, manual_shrink_to_fit);
 __turbo_tasks_internal_primitive!(RcStr);
 __turbo_tasks_internal_primitive!(Option<String>);
 __turbo_tasks_internal_primitive!(Option<RcStr>);
-__turbo_tasks_internal_primitive!(Vec<RcStr>);
+__turbo_tasks_internal_primitive!(Vec<RcStr>, manual_shrink_to_fit);
 __turbo_tasks_internal_primitive!(Option<u16>);
 __turbo_tasks_internal_primitive!(Option<u64>);
 __turbo_tasks_internal_primitive!(bool);
@@ -30,8 +30,8 @@ __turbo_tasks_internal_primitive!(usize);
 __turbo_tasks_internal_primitive!(isize);
 __turbo_tasks_internal_primitive!(serde_json::Value);
 __turbo_tasks_internal_primitive!(Duration);
-__turbo_tasks_internal_primitive!(Vec<u8>);
-__turbo_tasks_internal_primitive!(Vec<bool>);
+__turbo_tasks_internal_primitive!(Vec<u8>, manual_shrink_to_fit);
+__turbo_tasks_internal_primitive!(Vec<bool>, manual_shrink_to_fit);
 
 #[turbo_tasks::value(transparent, eq = "manual")]
 #[derive(Debug, Clone)]